requirements.txt
numpy
pandas
scikit-learn
tqdm
gradio
langchain
langchain-core
langgraph
langchain_openai
python-dotenv